backend/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
sqlalchemy==2.0.48
pydantic==2.12.5
python-dotenv==1.2.2
python-multipart==0.0.22
pandas==3.0.1
numpy==2.4.3
psycopg2-binary==2.9.11
scipy==1.15.2
